Begin deprecating the metadata factory pattern.
This lays the groundwork for a simpler way to produce priority and
predicate metadata objects. The goal is to get rid of registered
producer factories.

The existing implementation of priority and predicate metadata uses a
heavy amount of abstraction and dependency injection which is not adding
value to the codebase. To wit, PredicateMetadata is an interface with
exactly one implementation, and "PriorityMetadata" is currently defined
as `interface{}`. Users of these types invariably use a type assertion
to get at the useful data.